rename collisions to collision-resolver
